Changed CalcScreenCoord structure due to integer type too small for NSpire screen size. Fixed bug. Need to implement an RLE decompressor for screen shot.
git-svn-id: http://svn.tilp.info/repos/tilp@3950

+++ libticalcs/trunk/src/ticalcs.h
@@ -290,11 +290,11 @@ typedef struct
 {
 	int		format;
 
-	uint8_t width;
-	uint8_t height;
+	unsigned int width;
+	unsigned int height;
 
-	uint8_t clipped_width;
-	uint8_t clipped_height;
+	unsigned int clipped_width;
+	unsigned int clipped_height;
 } CalcScreenCoord;
 
 /**
